+++ src/mathpaper/document/__init__.py
@@ -1,1 +1,14 @@
-"""Document module for mathpaper."""
+from mathpaper.document.test import Test
+from mathpaper.document.problem import Problem, MultipartProblem
+from mathpaper.document.parts import Part
+from mathpaper.document.layout import AnswerSpace, PartsGrid, SideFigure
+
+__all__ = [
+    "Test",
+    "Problem",
+    "MultipartProblem",
+    "Part",
+    "AnswerSpace",
+    "PartsGrid",
+    "SideFigure",
+]