+++ src/mathpaper/render/escaping.py
@@ -1,1 +1,17 @@
-"""Character escaping utilities."""
+_TYPST_SPECIAL = str.maketrans({
+    "\\": "\\\\",
+    "#":  "\\#",
+    "@":  "\\@",
+    "<":  "\\<",
+    ">":  "\\>",
+    "*":  "\\*",
+    "_":  "\\_",
+    "`":  "\\`",
+    "~":  "\\~",
+    "[":  "\\[",
+    "]":  "\\]",
+})
+
+
+def escape_typst_text(text: str) -> str:
+    return text.translate(_TYPST_SPECIAL)